break activity dialog into own component
