+++ web/src/containers/activity/EntryDetails.js
@@ -1,17 +1,11 @@
 import { Button, Review, Dialog } from '@cmsgov/design-system-core';
 import PropTypes from 'prop-types';
-import React, { Fragment, useMemo, useRef, useState } from 'react';
+import React, { useMemo, useRef, useState } from 'react';
 import { connect } from 'react-redux';
 
 import { selectActivityByIndex } from '../../reducers/activities.selectors';
+import ActivityDialog from './EntryDetailsDialog.js';
 
-import ContractorResources from './ContractorResources';
-import CostAllocate from './CostAllocate';
-import Costs from './Costs';
-import Overview from './Overview';
-import Goals from './Goals';
-import Schedule from './Schedule';
-import StandardsAndConditions from './StandardsAndConditions';
 import { t } from '../../i18n';
 
 const makeTitle = ({ name, fundingSource }, i) => {
@@ -61,44 +55,6 @@ const EntryDetails = ({ activityIndex, fundingSource, activityKey, name }) => {
     [collapsed]
   );
 
-  const titleElement = (
-    <Fragment>
-      <span>{title}</span>
-      <nav className="dialog--nav">
-        <ul className="ds-c-list ds-c-list--bare">
-          <li className="active">
-            <a href={`#activity-overview-${activityIndex}`}>
-              Activity overview
-            </a>
-          </li>
-          <li>
-            <a href={`#activity-goals-${activityIndex}`}>Goals</a>
-          </li>
-          <li>
-            <a href={`#activity-cost-categories-${activityIndex}`}>
-              In-house cost categories
-            </a>
-          </li>
-          <li>
-            <a href={`#activity-contractor-costs-${activityIndex}`}>
-              Private contractor costs
-            </a>
-          </li>
-          <li>
-            <a href={`#activity-cost-allocation-${activityIndex}`}>
-              Cost allocation
-            </a>
-          </li>
-          <li>
-            <a href={`#activity-ffp-${activityIndex}`}>
-              FFP and cost allocation
-            </a>
-          </li>
-        </ul>
-      </nav>
-    </Fragment>
-  );
-
   return (
     <div
       id={`activity-${activityKey}`}
@@ -113,26 +69,7 @@ const EntryDetails = ({ activityIndex, fundingSource, activityKey, name }) => {
         ]}
       </Review>
       {showModal && (
-        <Dialog
-          actions={[
-            <Button variation="primary" onClick={() => setShowModal(false)}>
-              Done
-            </Button>
-          ]}
-          ariaCloseLabel={`Close modal for ${title}`}
-          className="ds-c-dialog--full"
-          closeButtonVariation="transparent"
-          onExit={() => setShowModal(false)}
-          title={titleElement}
-        >
-          <Overview activityIndex={activityIndex} />
-          <Goals activityIndex={activityIndex} />
-          <Schedule activityIndex={activityIndex} />
-          <Costs activityIndex={activityIndex} />
-          <ContractorResources activityIndex={activityIndex} />
-          <CostAllocate activityIndex={activityIndex} />
-          <StandardsAndConditions activityIndex={activityIndex} />
-        </Dialog>
+        <ActivityDialog title={title} activityIndex={activityIndex} />
       )}
     </div>
   );
